services/vectorization-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
kafka-python==2.0.2
openai==1.51.0
httpx==0.27.2
langchain==0.3.7
langchain-openai==0.2.0
langchain-community==0.3.5
chromadb>=1.0.0
pypdf2==3.0.1
python-multipart==0.0.12
aiofiles==24.1.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==5.0.0
httpx==0.27.2
